--- src/main/java/student/management/StudentManagement/StudentCourses.java
@@ -1,18 +1,0 @@
-package student.management.StudentManagement;
-
-import lombok.Getter;
-import lombok.Setter;
-@Getter
-@Setter
-
-public class StudentCourses {
-
-
-    private String id;
-    private String student_id;
-    private String course_name;
-    private String course_start_at;
-    private String course_end_at;
-
-
-}
