+++ src/main/java/student/management/StudentManagement/data/StudentCourses.java
@@ -1,0 +1,19 @@
+package student.management.StudentManagement.data;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+
+public class StudentCourses {
+
+
+  private String id;
+  private String student_id;
+  private String course_name;
+  private String course_start_at;
+  private String course_end_at;
+
+
+}
