+++ src/main/java/student/management/StudentManagement/service/StudentService.java
@@ -1,0 +1,27 @@
+package student.management.StudentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import student.management.StudentManagement.data.Student;
+import student.management.StudentManagement.data.StudentCourses;
+import student.management.StudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    return repository.search();
+  }
+
+  public List<StudentCourses> searchStudentCourseList() {
+    return repository.search2();
+  }
+}
